augurpy to class
