+++ .flake8
@@ -1,6 +1,6 @@
 [flake8]
 select = B,B9,C,E,F,N,W
-ignore = E203,E501,RST201,RST203,RST301,W503,D100,F401
+ignore = E203,E501,RST201,RST203,RST301,W503,D100,F401,B950
 max-line-length = 120
 max-complexity = 15
 docstring-convention = google
