+++ docs/code_of_conduct.md
@@ -1,2 +1,3 @@
 ```{include} ../CODE_OF_CONDUCT.md
+
 ```
