--- docs/usage/differential_prioritization/pertpy.tools.augurpy.differential_prioritization.predict_differential_prioritization.rst
@@ -1,6 +1,0 @@
-﻿pertpy.tools.augurpy.differential\_prioritization.predict\_differential\_prioritization
-=======================================================================================
-
-.. currentmodule:: pertpy.tools.augurpy.differential_prioritization
-
-.. autofunction:: predict_differential_prioritization
