--- docs/usage/estimator/pertpy.tools.augurpy.estimator.Params.rst
@@ -1,6 +1,0 @@
-﻿pertpy.tools.augurpy.estimator.Params
-=====================================
-
-.. currentmodule:: pertpy.tools.augurpy.estimator
-
-.. autoclass:: Params
