--- docs/usage/estimator/pertpy.tools.augurpy.estimator.create_estimator.rst
@@ -1,6 +1,0 @@
-﻿pertpy.tools.augurpy.estimator.create\_estimator
-================================================
-
-.. currentmodule:: pertpy.tools.augurpy.estimator
-
-.. autofunction:: create_estimator
