--- docs/usage/evaluate/pertpy.tools.augurpy.evaluate.predict.rst
@@ -1,6 +1,0 @@
-﻿pertpy.tools.augurpy.evaluate.predict
-=====================================
-
-.. currentmodule:: pertpy.tools.augurpy.evaluate
-
-.. autofunction:: predict
