--- docs/usage/read_load/pertpy.tools.augurpy.read_load.load.rst
@@ -1,6 +1,0 @@
-﻿pertpy.tools.augurpy.read\_load.load
-====================================
-
-.. currentmodule:: pertpy.tools.augurpy.read_load
-
-.. autofunction:: load
