+++ noxfile.py
@@ -105,7 +105,7 @@ def precommit(session: Session) -> None:
     session.install(
         "black",
         "flake8",
-        "flake8-bandit",
+        # "flake8-bandit",
         "flake8-bugbear",
         "flake8-docstrings",
         "flake8-rst-docstrings",
